split axissvr module
